refactor: 💡 make entities/Account folder, move MultSig into it
create a folder to better manage the complexity of Account. Move
MultiSig under Account, as it is a sub type
