+++ src/api/entities/Account/MultiSig/types.ts
@@ -1,0 +1,8 @@
+import BigNumber from 'bignumber.js';
+
+import { Signer } from '~/types';
+
+export interface MultiSigDetails {
+  signers: Signer[];
+  requiredSignatures: BigNumber;
+}
