+++ src/api/entities/types.ts
@@ -52,6 +52,6 @@ export * from './Offering/types';
 export * from './TickerReservation/types';
 export * from './Venue/types';
 export * from './Subsidy/types';
-export * from './MultiSig/types';
+export * from './Account/MultiSig/types';
 export * from './MultiSigProposal/types';
 export * from './MetadataEntry/types';
